+++ .coveragerc
@@ -1,3 +1,16 @@
+[paths]
+source = src
+
 [run]
+branch = true
+source =
+    src
+    tests
+parallel = true
 omit =
-    *custo*
+    *custom*
+
+[report]
+show_missing = true
+precision = 2
+omit = *migrations*
