+++ tests/solph_tests.py
@@ -11,8 +11,8 @@ SPDX-License-Identifier: MIT
 
 from nose.tools import ok_
 
-from oemof.energy_system import EnergySystem as EnSys
-from oemof.network import Node
+from oemof.network.energy_system import EnergySystem as EnSys
+from oemof.network.network import Node
 from oemof.solph.blocks import InvestmentFlow as InvFlow
 from oemof.solph import Investment
 import oemof.solph as solph
