+++ tests/test_network_classes.py
@@ -12,9 +12,9 @@ SPDX-License-Identifier: MIT
 from traceback import format_exception_only as feo
 from nose.tools import assert_raises, eq_, ok_
 
-from oemof.energy_system import EnergySystem as EnSys
-from oemof.network import (Bus, Edge, Node, Transformer, registry_changed_to,
-                           temporarily_modifies_registry)
+from oemof.network.energy_system import EnergySystem as EnSys
+from oemof.network.network import (Bus, Edge, Node, Transformer, registry_changed_to,
+                                   temporarily_modifies_registry)
 
 
 class TestsNode:
