+++ tests/test_scripts/test_solph/test_simple_model/test_simple_invest.py
@@ -19,9 +19,9 @@ import pandas as pd
 from oemof.solph import (Sink, Source, Transformer, Bus, Flow, Model,
                          EnergySystem, Investment)
 from oemof.tools import economics
-from oemof.outputlib import views
+from oemof.network import views
 from oemof.solph import processing
-from oemof.network import Node
+from oemof.network.network import Node
 
 
 def test_dispatch_example(solver='cbc', periods=24*5):
