+++ tests/test_scripts/test_solph/test_storage_investment/test_storage_investment.py
@@ -41,8 +41,8 @@ from nose.tools import eq_
 from oemof.tools import economics
 
 import oemof.solph as solph
-from oemof.network import Node
-from oemof.outputlib import views
+from oemof.network.network import Node
+from oemof.network import views
 from oemof.solph import processing
 
 import logging
