+++ tests/test_scripts/test_solph/test_variable_chp/test_variable_chp.py
@@ -19,7 +19,7 @@ import pandas as pd
 
 from oemof import outputlib
 
-from oemof.network import Node
+from oemof.network.network import Node
 import oemof.solph as solph
