+++ tests/test_warnings.py
@@ -11,7 +11,8 @@ SPDX-License-Identifier: MIT
 
 from nose.tools import eq_, ok_, with_setup
 import warnings
-from oemof import network, solph
+from oemof import solph
+from oemof.network import network
 from oemof.tools.debugging import SuspiciousUsageWarning
